Move deprecated components to src/deprecated
