+++ src/deprecated/SelectMenu/SelectMenu.tsx
@@ -1,7 +1,7 @@
 import React, {useCallback, useEffect, useRef, useState} from 'react'
 import styled from 'styled-components'
-import sx, {SxProp} from '../sx'
-import {ComponentProps} from '../utils/types'
+import sx, {SxProp} from '../../sx'
+import {ComponentProps} from '../../utils/types'
 import useKeyboardNav from './hooks/useKeyboardNav'
 import {MenuContext} from './SelectMenuContext'
 import SelectMenuDivider from './SelectMenuDivider'
