+++ src/deprecated/SelectMenu/SelectMenuFilter.tsx
@@ -1,9 +1,9 @@
 import React, {forwardRef, useContext, useEffect, useRef} from 'react'
 import styled from 'styled-components'
-import {get} from '../constants'
-import sx, {SxProp} from '../sx'
-import TextInput, {TextInputProps} from '../TextInput'
-import {ComponentProps} from '../utils/types'
+import {get} from '../../constants'
+import sx, {SxProp} from '../../sx'
+import TextInput, {TextInputProps} from '../../TextInput'
+import {ComponentProps} from '../../utils/types'
 import {MenuContext} from './SelectMenuContext'
 
 const StyledForm = styled.form<SxProp>`
