+++ src/deprecated/SelectMenu/SelectMenuLoadingAnimation.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
-import {get} from '../constants'
-import Spinner from '../Spinner'
-import sx, {SxProp} from '../sx'
-import {ComponentProps} from '../utils/types'
+import {get} from '../../constants'
+import Spinner from '../../Spinner'
+import sx, {SxProp} from '../../sx'
+import {ComponentProps} from '../../utils/types'
 
 const Animation = styled.div<SxProp>`
   padding: ${get('space.6')} ${get('space.4')};
