+++ src/deprecated/index.ts
@@ -6,30 +6,30 @@
  *  example: import {FormGroup} from '@primer/react/deprecated
  */
 
-export {default as BorderBox} from '../BorderBox'
-export type {BorderBoxProps} from '../BorderBox'
+export {default as BorderBox} from './BorderBox'
+export type {BorderBoxProps} from './BorderBox'
 export {default as ChoiceFieldset, Item} from './ChoiceFieldset'
 export {default as ChoiceInputField} from './ChoiceInputField'
-export {default as Flex} from '../Flex'
-export type {FlexProps} from '../Flex'
-export {default as Grid} from '../Grid'
-export type {GridProps} from '../Grid'
-export {default as Position, Absolute, Fixed, Relative, Sticky} from '../Position'
-export type {PositionProps, AbsoluteProps, FixedProps, RelativeProps, StickyProps} from '../Position'
-export {default as Dropdown} from '../Dropdown'
+export {default as Flex} from './Flex'
+export type {FlexProps} from './Flex'
+export {default as Grid} from './Grid'
+export type {GridProps} from './Grid'
+export {default as Position, Absolute, Fixed, Relative, Sticky} from './Position'
+export type {PositionProps, AbsoluteProps, FixedProps, RelativeProps, StickyProps} from './Position'
+export {default as Dropdown} from './Dropdown'
 export type {
   DropdownProps,
   DropdownCaretProps,
   DropdownButtonProps,
   DropdownItemProps,
   DropdownMenuProps
-} from '../Dropdown'
+} from './Dropdown'
 export {default as FormGroup} from './FormGroup'
 export type {FormGroupProps, FormGroupLabelProps} from './FormGroup'
 export {default as InputField} from './InputField'
 export {default as Label} from './Label'
 export type {LabelProps} from './Label'
-export {default as SelectMenu} from '../SelectMenu'
+export {default as SelectMenu} from './SelectMenu'
 export type {
   SelectMenuProps,
   SelectMenuDividerProps,
@@ -43,7 +43,7 @@ export type {
   SelectMenuTabProps,
   SelectMenuTabPanelProps,
   SelectMenuLoadingAnimationProps
-} from '../SelectMenu'
+} from './SelectMenu'
 export {ActionList} from './ActionList'
 export type {ActionListProps} from './ActionList'
 export {ActionMenu} from './ActionMenu'
